+++ sdk/communication/azure-communication-identity/tests/test_communication_identity_client_async.py
@@ -5,6 +5,7 @@
 # license information.
 # --------------------------------------------------------------------------
 import pytest
+import datetime
 from azure.core.credentials import AccessToken
 from azure.communication.identity.aio import CommunicationIdentityClient
 from azure.communication.identity import CommunicationTokenScope
@@ -30,7 +31,7 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
             BodyReplacerProcessor(keys=["id", "token"]),
             URIIdentityReplacer()])
     
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @ResourceGroupPreparer(random_name_enabled=True, delete_after_tag_timedelta=datetime.timedelta(hours=2))
     @CommunicationServicePreparer()
     async def test_create_user_from_managed_identity(self, connection_string):
         endpoint, access_key = parse_connection_str(connection_string)
@@ -43,28 +44,28 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
         async with identity_client:
             user = await identity_client.create_user()
 
-        assert user.identifier is not None
+        assert user.properties.get('id') is not None
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @ResourceGroupPreparer(random_name_enabled=True, delete_after_tag_timedelta=datetime.timedelta(hours=2))
     @CommunicationServicePreparer()
     async def test_create_user(self, connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(connection_string)
         async with identity_client:
             user = await identity_client.create_user()
 
-        assert user.identifier is not None
+        assert user.properties.get('id') is not None
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @ResourceGroupPreparer(random_name_enabled=True, delete_after_tag_timedelta=datetime.timedelta(hours=2))
     @CommunicationServicePreparer()
     async def test_create_user_and_token(self, connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(connection_string)
         async with identity_client:
             user, token_response = await identity_client.create_user_and_token(scopes=[CommunicationTokenScope.CHAT])
 
-        assert user.identifier is not None
+        assert user.properties.get('id') is not None
         assert token_response.token is not None
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @ResourceGroupPreparer(random_name_enabled=True, delete_after_tag_timedelta=datetime.timedelta(hours=2))
     @CommunicationServicePreparer()
     async def test_get_token_from_managed_identity(self, connection_string):
         endpoint, access_key = parse_connection_str(connection_string)
@@ -78,10 +79,10 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
             user = await identity_client.create_user()
             token_response = await identity_client.get_token(user, scopes=[CommunicationTokenScope.CHAT])
 
-        assert user.identifier is not None
+        assert user.properties.get('id') is not None
         assert token_response.token is not None
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @ResourceGroupPreparer(random_name_enabled=True, delete_after_tag_timedelta=datetime.timedelta(hours=2))
     @CommunicationServicePreparer()
     async def test_get_token(self, connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(connection_string)
@@ -89,10 +90,10 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
             user = await identity_client.create_user()
             token_response = await identity_client.get_token(user, scopes=[CommunicationTokenScope.CHAT])
 
-        assert user.identifier is not None
+        assert user.properties.get('id') is not None
         assert token_response.token is not None
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @ResourceGroupPreparer(random_name_enabled=True, delete_after_tag_timedelta=datetime.timedelta(hours=2))
     @CommunicationServicePreparer()
     async def test_revoke_tokens_from_managed_identity(self, connection_string):
         endpoint, access_key = parse_connection_str(connection_string)
@@ -107,10 +108,10 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
             token_response = await identity_client.get_token(user, scopes=[CommunicationTokenScope.CHAT])
             await identity_client.revoke_tokens(user)
 
-        assert user.identifier is not None
+        assert user.properties.get('id') is not None
         assert token_response.token is not None
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @ResourceGroupPreparer(random_name_enabled=True, delete_after_tag_timedelta=datetime.timedelta(hours=2))
     @CommunicationServicePreparer()
     async def test_revoke_tokens(self, connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(connection_string)
@@ -119,10 +120,10 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
             token_response = await identity_client.get_token(user, scopes=[CommunicationTokenScope.CHAT])
             await identity_client.revoke_tokens(user)
 
-        assert user.identifier is not None
+        assert user.properties.get('id') is not None
         assert token_response.token is not None
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @ResourceGroupPreparer(random_name_enabled=True, delete_after_tag_timedelta=datetime.timedelta(hours=2))
     @CommunicationServicePreparer()
     async def test_delete_user_from_managed_identity(self, connection_string):
         endpoint, access_key = parse_connection_str(connection_string)
@@ -136,9 +137,9 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
             user = await identity_client.create_user()
             await identity_client.delete_user(user)
 
-        assert user.identifier is not None
+        assert user.properties.get('id') is not None
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @ResourceGroupPreparer(random_name_enabled=True, delete_after_tag_timedelta=datetime.timedelta(hours=2))
     @CommunicationServicePreparer()
     async def test_delete_user(self, connection_string):
         identity_client = CommunicationIdentityClient.from_connection_string(connection_string)
@@ -146,4 +147,4 @@ class CommunicationIdentityClientTestAsync(AsyncCommunicationTestCase):
             user = await identity_client.create_user()
             await identity_client.delete_user(user)
 
-        assert user.identifier is not None
+        assert user.properties.get('id') is not None