+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_id_documents_from_url_async.py
@@ -63,7 +63,7 @@ class TestIdDocumentsFromUrlAsync(AsyncFormRecognizerTest):
     @FormRecognizerPreparer()
     @GlobalClientPreparer()
     async def test_id_document_url_pass_stream(self, client):
-        with open(self.id_document_jpg, "rb") as id_document:
+        with open(self.id_document_license_jpg, "rb") as id_document:
             with self.assertRaises(HttpResponseError):
                 async with client:
                     poller = await client.begin_recognize_id_documents_from_url(id_document)
@@ -103,6 +103,27 @@ class TestIdDocumentsFromUrlAsync(AsyncFormRecognizerTest):
 
         # Check page metadata
         self.assertFormPagesTransformCorrect(id_document.pages, read_results, page_results)
+
+    @FormRecognizerPreparer()
+    @GlobalClientPreparer()
+    async def test_id_document_jpg_passport(self, client):
+        async with client:
+            poller = await client.begin_recognize_id_documents_from_url(self.id_document_url_jpg_passport)
+
+            result = await poller.result()
+            self.assertEqual(len(result), 1)
+        
+            id_document = result[0]
+            # check dict values
+
+            passport = id_document.fields.get("MachineReadableZone").value
+            self.assertEqual(passport["LastName"].value, "MARTIN")
+            self.assertEqual(passport["FirstName"].value, "SARAH")
+            self.assertEqual(passport["DocumentNumber"].value, "ZE000509")
+            self.assertEqual(passport["DateOfBirth"].value, date(1985,1,1))
+            self.assertEqual(passport["DateOfExpiration"].value, date(2023,1,14))
+            self.assertEqual(passport["Sex"].value, "F")
+            self.assertEqual(passport["Country"].value, "CAN")
 
     @FormRecognizerPreparer()
     @GlobalClientPreparer()