+++ sdk/tables/azure-data-tables/tests/test_table_entity_cosmos_async.py
@@ -1625,7 +1625,7 @@ class StorageTableEntityTest(AzureTestCase, AsyncTableTestCase):
                 sleep(SLEEP_DELAY)
 
     @CosmosPreparer()
-    async def test_query_injection(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
+    async def test_query_injection_async(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
         await self._set_up(tables_cosmos_account_name, tables_primary_cosmos_account_key)
         try:
@@ -1657,8 +1657,8 @@ class StorageTableEntityTest(AzureTestCase, AsyncTableTestCase):
                 entities.append(e)
             assert len(entities) ==  0
         finally:
-            self.ts.delete_table(table_name)
-            self._tear_down()
+            await self.ts.delete_table(table_name)
+            await self._tear_down()
 
     @CosmosPreparer()
     async def test_query_special_chars(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
@@ -2083,8 +2083,6 @@ class StorageTableEntityTest(AzureTestCase, AsyncTableTestCase):
 
     @CosmosPreparer()
     async def test_datetime_milliseconds(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
-        # SAS URL is calculated from storage key, so this test runs live only
-        url = self.account_url(tables_cosmos_account_name, "table")
         await self._set_up(tables_cosmos_account_name, tables_primary_cosmos_account_key)
         try:
             entity = self._create_random_entity_dict()
@@ -2103,3 +2101,31 @@ class StorageTableEntityTest(AzureTestCase, AsyncTableTestCase):
         finally:
             await self._tear_down()
             self.sleep(SLEEP_DELAY)
+
+    @CosmosPreparer()
+    async def test_datetime_str_passthrough(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
+        await self._set_up(tables_cosmos_account_name, tables_primary_cosmos_account_key)
+        partition, row = self._create_pk_rk(None, None)
+
+        dotnet_timestamp = "2013-08-22T01:12:06.2608595Z"
+        entity = {
+            'PartitionKey': partition,
+            'RowKey': row,
+            'datetime1': EntityProperty(dotnet_timestamp, EdmType.DATETIME)
+        }
+        try:
+            await self.table.create_entity(entity)
+            received = await self.table.get_entity(partition, row)
+            assert isinstance(received['datetime1'], datetime)
+            assert received.datetime1.tables_service_value == dotnet_timestamp
+
+            received['datetime2'] = received.datetime1.replace(year=2020)
+            assert received['datetime2'].tables_service_value == ""
+
+            await self.table.update_entity(received, mode=UpdateMode.REPLACE)
+            updated = await self.table.get_entity(partition, row)
+            assert isinstance(updated['datetime1'], datetime)
+            assert isinstance(updated['datetime2'], datetime)
+            assert updated.datetime1.tables_service_value == dotnet_timestamp
+        finally:
+            await self._tear_down()