+++ sdk/containerregistry/azure-containerregistry/CHANGELOG.md
@@ -1,4 +1,4 @@
 # Release History
 
-## 0.0.1 (Unreleased)
+## 1.0.0b1 (2021-04-06)
 * First release of the Azure Container Registry library for Python