+++ sdk/core/azure-core/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 1.13.1 (Unreleased)
+
+
 ## 1.13.0 (2021-04-02)
 
 Azure core requires Python 2.7 or Python 3.6+ since this release.
@@ -8,6 +11,8 @@ Azure core requires Python 2.7 or Python 3.6+ since this release.
 
 - Added `azure.core.utils.parse_connection_string` function to parse connection strings across SDKs, with common validation and support for case insensitive keys.
 - Supported adding custom policies  #16519
+- Added `~azure.core.tracing.Link` that should be used while passing `Links` to  `AbstractSpan`.
+- `AbstractSpan` constructor can now take in additional keyword only args.
 
 ### Bug fixes